Split some implementation details out from the api module in order to support windup-web
