+++ config/api/src/main/java/org/jboss/windup/config/metadata/TechnologyMetadataProvider.java
@@ -1,26 +1,15 @@
 package org.jboss.windup.config.metadata;
 
-import javax.inject.Inject;
-
-import org.jboss.forge.furnace.services.Imported;
 import org.jboss.windup.graph.GraphContext;
 
 /**
+ *
+ * Maintains information about which technologies are defined in the rulesets.
+ *
  * @author <a href="mailto:jesse.sightler@gmail.com">Jesse Sightler</a>
  */
-public class TechnologyMetadataProvider
-{
-    @Inject
-    private Imported<TechnologyMetadataLoader> loaders;
 
-    public TechnologyMetadata getMetadata(GraphContext context, TechnologyReference reference)
-    {
-        for (TechnologyMetadataLoader loader : loaders)
-        {
-            TechnologyMetadata metadata = loader.getMetadata(context, reference);
-            if (metadata != null)
-                return metadata;
-        }
-        return null;
-    }
+public interface TechnologyMetadataProvider
+{
+    TechnologyMetadata getMetadata(GraphContext context, TechnologyReference reference);
 }
